internal/log/segment.go
package log

import (
	"fmt"
	"github.com/golang/protobuf/proto"
	api "github.com/vsivarajah/distributed_services/api/v1"
	"os"
	"path"
)

type segment struct {
	store                  *store
	index                  *index
	baseOffset, nextOffset uint64
	config                 Config
}

//the log calls newSegment when it needs to add a new segment, such as when the current active segment hits its maxsize
//open the store and index files and pass the os.O_CREATE file mode flag as an argument to openfile to create the files if they don't exist yet.
func newSegment(dir string, baseOffset uint64, c Config) (*segment, error) {
	s := &segment{
		baseOffset: baseOffset,
		config:     c,
	}
	var err error
	storeFile, err := os.OpenFile(
		path.Join(dir, fmt.Sprintf("%d%s", baseOffset, ".store")),
		os.O_RDWR|os.O_CREATE|os.O_APPEND,
		0644,
	)
	if err != nil {
		return nil, err
	}
	if s.store, err = newStore(storeFile); err != nil {
		return nil, err
	}
	indexFile, err := os.OpenFile(
		path.Join(dir, fmt.Sprintf("%d%s", baseOffset, ".index")),
		os.O_RDWR|os.O_CREATE,
		0644,
	)

	if err != nil {
		return nil, err
	}
	if s.index, err = newIndex(indexFile, c); err != nil {
		return nil, err
	}
	if off, _, err := s.index.Read(-1); err != nil {
		//if index is empty, then the next reccord appended to the segment would be the first record and its offset would be the segment's baseoffset.
		s.nextOffset = baseOffset
	} else {
		//if the index has at least one entry, then the offset of the next record written take the offset at the end of the segment by adding 1
		//to the baseoffset and relative offset.
		s.nextOffset = baseOffset + uint64(off) + 1
	}
	return s, nil
}

//Append writes the record to the segment and returns the newly appended record's offset.
//The log returns the offset to the API response.
func (s *segment) Append(record *api.Record) (offset uint64, err error) {
	cur := s.nextOffset
	record.Offset = cur
	p, err := proto.Marshal(record)
	if err != nil {
		return 0, err
	}
	_, pos, err := s.store.Append(p)
	if err != nil {
		return 0, err
	}
	if err = s.index.Write(
		//index offsets are relative to base offset
		uint32(s.nextOffset-uint64(s.baseOffset)),
		pos,
	); err != nil {
		return 0, err
	}
	s.nextOffset++
	return cur, nil
}

//Read returns the record for the given offset. Similar to writes, to read a record
//the segment must first translate the absolute index into a relative offset
//and get the associated index entry.
//once it has the index entry, the segment can go straight to the record's position in the store and read the proper amount of data.
func (s *segment) Read(off uint64) (*api.Record, error) {
	_, pos, err := s.index.Read(int64(off - s.baseOffset))
	if err != nil {
		return nil, err
	}
	p, err := s.store.Read(pos)
	if err != nil {
		return nil, err
	}
	record := &api.Record{}
	err = proto.Unmarshal(p, record)
	return record, err
}

//IsMaxed returns whether the segment has reached its max size
//either by writing too much to the store or the index.
func (s *segment) IsMaxed() bool {
	return s.store.size >= s.config.Segment.MaxStoreBytes ||
		s.index.size >= s.config.Segment.MaxIndexBytes
}

//Remove closes the segment and removes the index and store files
func (s *segment) Remove() error {
	if err := s.Close(); err != nil {
		return err
	}
	if err := os.Remove(s.index.Name()); err != nil {
		return err
	}
	if err := os.Remove(s.store.Name()); err != nil {
		return err
	}
	return nil
}

func (s *segment) Close() error {
	if err := s.index.Close(); err != nil {
		return err
	}
	if err := s.store.Close(); err != nil {
		return err
	}
	return nil
}

//nearestMultiple returns the nearest and lesser multiple of k in j.
//ex: nearestMultiple(9,4) == 8.
//take the lesser multiple to make sure we stay under the user's disk capacity.
func nearestMultiple(j, k uint64) uint64 {
	if j >= 0 {
		return (j / k) * k
	}
	return ((j - k + 1) / k) * k
}
